move stuff into utils folder and create navigator
