+++ src/app/loginPage.jsx
@@ -5,19 +5,13 @@ import MainContent from './mainContent'
 
 import TitleBar from './titleBar'
 
-class LoginPage extends React.Component {
-
-  render() {
-    return (
-      <section>
-        <TitleBar />
-        <MainContent>
-          <LoginPanel />
-        </MainContent>
-      </section>
-    )
-  }
-
-}
+const LoginPage = () => (
+  <section>
+    <TitleBar />
+    <MainContent>
+      <LoginPanel />
+    </MainContent>
+  </section>
+)
 
 export default LoginPage
