--- src/app/setPage.js
@@ -1,5 +1,0 @@
-import state from './state/'
-
-export default function setPage(page) {
-  state.setPage(page)
-}
